Desenvolvimento Kotlin PCP VII - Chamado 22594

+++ app/src/main/java/br/com/usinasantafe/pcpk/features/domain/repositories/variable/MovEquipProprioSegRepository.kt
@@ -19,6 +19,6 @@ interface MovEquipProprioSegRepository {
 
     suspend fun listEquipSeg(idMov: Long): List<MovEquipProprioSeg>
 
-    suspend fun saveEquipSeg(idMov: Int): Boolean
+    suspend fun saveEquipSeg(idMov: Long): Boolean
 
 }
