+++ app/src/main/java/br/com/usinasantafe/pcpk/features/domain/repositories/variable/MovEquipResidenciaRepository.kt
@@ -15,9 +15,9 @@ interface MovEquipResidenciaRepository {
 
     suspend fun receiverSentMovEquipResidencia(movEquipList: List<MovEquipResidencia>): Boolean
 
-    suspend fun saveMovEquipResidencia(matricVigia: Long, idLocal: Long): Int
+    suspend fun saveMovEquipResidencia(matricVigia: Long, idLocal: Long): Long
 
-    suspend fun saveMovEquipResidencia(matricVigia: Long, idLocal: Long, movEquipResidencia: MovEquipResidencia): Int
+    suspend fun saveMovEquipResidencia(matricVigia: Long, idLocal: Long, movEquipResidencia: MovEquipResidencia): Long
 
     suspend fun sendMovEquipResidencia(movEquipList: List<MovEquipResidencia>, nroAparelho: Long): Result<List<MovEquipResidencia>>
