+++ app/src/main/java/br/com/usinasantafe/pcpk/features/infra/datasource/room/variable/MovEquipResidenciaDatasourceRoom.kt
@@ -10,7 +10,7 @@ interface MovEquipResidenciaDatasourceRoom {
 
     suspend fun insertMovEquipResidenciaClose(movEquipResidenciaRoomModel: MovEquipResidenciaRoomModel): Boolean
 
-    suspend fun lastIdMovStatusSend(): Int
+    suspend fun lastIdMovStatusSend(): Long
 
     suspend fun listMovEquipResidenciaOpen(): List<MovEquipResidenciaRoomModel>
